requirements.txt
# Core dependencies
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=8.0.0
requests>=2.25.0

# For API interactions
google-generativeai>=0.0.1

# For image processing
scikit-image>=0.18.0

# For data handling
pandas>=1.3.0

# For testing
pytest>=6.0.0
pytest-cov>=2.12.0

# For documentation
Sphinx>=4.0.0